SymbolicArrays

Documentation for SymbolicArrays.

SymbolicArrays.SymbolicArray

SymbolicArrays.SymbolicNamedDimArraySum

SymbolicArrays.SymbolicNamedDimArraySum

SymbolicArrays.arguments

SymbolicArrays.arguments_contract

SymbolicArrays.arguments_sum

SymbolicArrays.coefficient

SymbolicArrays.coefficient

SymbolicArrays.coefficients

SymbolicArrays.flatten_expression

SymbolicArrays.leaf_arguments

SymbolicArrays.operator

SymbolicArrays.space_complexity

SymbolicArrays.substitute

SymbolicArrays.SymbolicArray — Type

A symbolic array, without dimension names.

SymbolicArrays.SymbolicNamedDimArraySum — Method

Sum based on arguments.

SymbolicArrays.SymbolicNamedDimArraySum — Method

Sum based on arguments.

SymbolicArrays.arguments — Method

The (NamedDimArray) arguments of a NamedDimArray contraction or sum. See also SymbolicUtils.arguments and TermInterface.arguments.

SymbolicArrays.arguments_contract — Method

The (NamedDimArray) arguments of a NamedDimArray contraction.

SymbolicArrays.arguments_sum — Method

The (NamedDimArray) arguments of a NamedDimArray sum.

SymbolicArrays.coefficient — Method

The coefficient of an unscaled term in a sum.

SymbolicArrays.coefficient — Method

The coefficient of a NamedDimArray.

SymbolicArrays.coefficients — Method

A dictionary from unscaled arguments/terms of a sum to their corresponding coefficients.

SymbolicArrays.flatten_expression — Method

Flatten a nested expression down to a flat expression, removing information about the order of operations.

SymbolicArrays.leaf_arguments — Method

The leaf (NamedDimArray) arguments of a nested NamedDimArray contraction or sum.

SymbolicArrays.operator — Method

The function/operator associated with the head/root of the symbolic expression, see also SymbolicUtils.operator and arguments.

SymbolicArrays.space_complexity — Method

The space complexity of the expression.

SymbolicArrays.substitute — Method

Substitute the specified subexpression for a new one.