SymbolicArrays

Documentation for SymbolicArrays.

SymbolicArrays.SymbolicArray

SymbolicArrays.TensorExpr.Sum

SymbolicArrays.TensorExpr.Sum

SymbolicArrays.arguments

SymbolicArrays.arguments_contract

SymbolicArrays.arguments_sum

SymbolicArrays.coefficient

SymbolicArrays.coefficient

SymbolicArrays.coefficients

SymbolicArrays.leaf_arguments

SymbolicArrays.space_complexity

SymbolicArrays.substitute

SymbolicArrays.time_complexity

SymbolicArrays.SymbolicArray — Type

A symbolic array, without dimension names.

SymbolicArrays.TensorExpr.Sum — Method

Sum based on arguments.

SymbolicArrays.TensorExpr.Sum — Method

Sum based on arguments.

SymbolicArrays.arguments — Method

The (tensor) arguments of a tensor contraction or sum.

SymbolicArrays.arguments_contract — Method

The (tensor) arguments of a tensor contraction.

SymbolicArrays.arguments_sum — Method

The (tensor) arguments of a tensor sum.

SymbolicArrays.coefficient — Method

The coefficient of an unscaled term in a sum.

SymbolicArrays.coefficient — Method

The coefficient of a tensor.

SymbolicArrays.coefficients — Method

A dictionary from unscaled arguments/terms of a sum to their corresponding coefficients.

SymbolicArrays.leaf_arguments — Method

The leaf (tensor) arguments of a nested tensor contraction or sum.

SymbolicArrays.space_complexity — Method

The space complexity of the head/root expression.

To get the total complexity of the full expression, walk the expression tree. See: https://docs.juliahub.com/General/EinExprs/stable/counters https://github.com/JuliaCollections/AbstractTrees.jl https://github.com/chengchingwen/StructWalk.jl https://github.com/FluxML/Functors.jl

SymbolicArrays.substitute — Method

Substitute the specified subexpression for a new one.

SymbolicArrays.time_complexity — Method

The time complexity of the head/root expression.

To get the total complexity of the full expression, walk the expression tree. See: https://docs.juliahub.com/General/EinExprs/stable/counters https://github.com/JuliaCollections/AbstractTrees.jl https://github.com/chengchingwen/StructWalk.jl https://github.com/FluxML/Functors.jl